Fix cases of no Property generated in XML output.
Modify the JKind class to add an "empty" JKindResult in the event that no Property is generated in the results XML file (typically caused by Java exception in JKind tool).
Modify the JKindResult class to directly print errors and removed the internal failLog list. This in turn simplifies the Test Case. Also got rid of a nasty problem where stale failLogs fooled the tests in to erroneous failures.

Few other random cleanup and doc updates.